C/R: Fix usage of type long for the appid in commands checkpoint and restart
Port of SVN trunk r5227

The application identifier is a long that is initialized to the pid of the first
process creating the application. The process identifier is of type pid_t, that
is a int. Thus, *currently*, a int is always big enough to store the application
identifier.

Anyway, the way we choose application identifier may change, so it is
safer to fix the type in checkpoint/restart commands.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>