Implement C/R callbacks
Port of SVN trunk r5311

Merge C/R callbacks from branch cb_checkpoint.

Callbacks are always executed before taking a checkpoint and during the restart
in order to be able to save/restore resources which the kernel checkpointer is
not capable off.

In order to use the callback library each application must be linked against it.
Furthermore the master process of the application is in charge for calling the
init function in order to initialize the callback library. The default behavior
is that the master process of the application is in charge of callback
execution. After the initialization the registration of callbacks must be done.

A signal has to be send by the checkpoint/restart command to the master process
in order to trigger the callback execution. We employ three free signals (37, 38
and 39) and one message queue in order to realize the communication between the
callback library and the checkpoint/restart commands. The first signal is used
to tell the process to execute callbacks on checkpoint. The second signal is
used to tell the process to execute callbacks on restart. The third signal is
used to execute the continuing callbacks. Continuing callbacks can be used to
recreate resources after a successful checkpoint without killing and restarting
the application explicitly.

We use a message queue to receive the answer from
the callback library about the status (error or no) of the callback execution
out of the checkpoint command.

Since there is currently no way to checkpoint and restart pipes, sockets, files
appropriate we have decided to use signals in order realize the communication
between the callback library and the checkpoint/restart commands. The message
queue is used temporary to received the status from the callback library and
deleted shortly before taking a checkpoint.

Another possible approach for callback registration:

It is possible for the programmer to make the callback registration out of each
application process. Hence the cr_init_callback function must be called out of
each process in order to initialize the callback library appropriate. Afterwards
it is possible to register callbacks for the process as described above. In that
case the master process is in charge of sending the signals to the child process
to execute callbacks.

Signed-off-by: Eugen Feller <eugen.feller@uni-duesseldorf.de>
Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>