krg_ns: Introduce krginit_helper
krginit_helper is meant to be the temporary cluster init helper that completes
the setup of a local instance of the Kerrighed container before actually adding
the local node to a cluster. This last step is done by exec()'ing krginit at the
end of the setup.

krginit_helper has its configuration in /etc/kerrighed/krginit_helper.conf

krginit_helper can also be used on the node starting the cluster, and will
prefer arguments passed by the caller to its configuration file.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>