krg_ns: Introduce the krgboot helper
This commit introduces the krgboot helper as a wrapper setting a program creating a
container as a Kerrighed cluster creator.

krgboot also accepts options telling which namespaces to isolate the Kerrighed
container from. In that case krgboot will create itself the Kerrighed container
and execute the helper program to finish the container's setup.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>