krgadm nodes: replace option --auto by --count and add option --all
Option --auto of krgadm nodes add/del was taking as argument the number
of node for which the action should apply. The name of this option does
not reflect the fact that an argument is needed. Replace it by 'count'.

Add option --all that tells the action should be done for all available
nodes.

This patch fixes the misbehavior of the command when using for instance
'krgadm nodes del <nodeid>'. While this command is wrong (option -n is
missing), the result was somewhat surprising since it was trying to
remove all the nodes. With this patch, it displays help and exits.