C/R: Remove function krg_check_checkpoint()
krg_check_checkpoint() was only used to check if /var/chkpt exists. The
problem is that it prevent to get a clear message in case of failure
because the dirname gets hidden.

The function may be reintroduced one day if there is a way to check if
the kernel supports checkpointing.